Created a lightweight Scaffold object, render override support & more
There is a lightweight Scaffold object that is used to track
scaffolds and provide default functionality that can be overridden.

`render` - Now you can override the render method to disable
templating or to provide a special handler of your own.

`processData` - A new feature to allow you prep the data that is
passed to the templates after the user input has been received.

Closes #4. Partial work for #2. Partial work for #1